pyproject.toml
[tool.poetry]
name = "restful-booker-api-tests"
version = "0.1.0"
description = ""
authors = ["danilaKazimir <kazimird99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.32.3"
pydantic = "^2.10.6"
pytest = "^8.3.4"
faker = "^35.0.0"
deepdiff = "^8.2.0"
allure-pytest = "^2.13.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
